Move `config.rs` and `theme.rs` into `config` folder
